claude-vignette/requirements.txt
anthropic>=0.15.0
flask>=2.0.0
python-dotenv>=0.19.0
Pillow>=9.0.0